+++ src/application/mod.rs
@@ -61,6 +61,7 @@ pub mod importing_messages;
 pub mod invitations;
 pub mod item_fields;
 pub mod keeping_message_text;
+pub mod links_in_text;
 pub mod local_delete;
 pub mod local_folders;
 pub mod long_text;